Déplacement des stats sur la page d'accueil
