+++ ui/src/accueil/stats/NouveauxHistogram.jsx
